fix room error for return type

+++ Source/app/src/main/java/mb/games/loveletter/data/GameSession.kt
@@ -5,7 +5,7 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "game_sessions")
 data class GameSession(
-    @PrimaryKey(autoGenerate = true) val id: Int = 0,
+    @PrimaryKey(autoGenerate = true) val id: Long = 0L,
     val currentTurn: Int,
     val playerIds: List<Int>,
     val deckState: String
